requirements.txt
opencv-python
deepface 
langchain
langchain-groq
streamlit 
pillow 
numpy
tf-keras
langchain-community